+++ lib/base_state_notifier.dart
@@ -1,2 +1,3 @@
+export 'src/domain/notifiers/base_notifier.dart';
 export 'src/domain/notifiers/base_state.dart';
 export 'src/domain/notifiers/base_state_notifier.dart';
